+++ src/Symfony/Component/Form/CHANGELOG.md
@@ -133,3 +133,8 @@ CHANGELOG
    FormEvents::BIND_NORM_DATA
  * [BC BREAK] reversed the order of the first two arguments to `createNamed`
    and `createNamedBuilder` in `FormFactoryInterface`
+ * [BC BREAK] adapted methods of FormView to match the naming used in
+   FormInterface and FormBuilder
+ * deprecated `getChildren` in Form and FormBuilder in favor of `all`
+ * deprecated `hasChildren` in Form and FormBuilder in favor of `count`
+ * FormBuilder now implements \IteratorAggregate